Night 5 Commit
Night 5 Commit.

Corrected an error in that I didn't need to use a strategy pattern on
the DAOs.  THe Items end up needing to know all of their costs and
prices during various operations.  Its easier, in turn to pull this
strategy into the transaction logic its self.

The add item window pulls double-function right now as a lookup by UPC.
Should be able to extend the functionality into a proper buy window.